backend/requirements/prod.txt
-r common.txt
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0